Extract über/main fuzzer into a module
I want to reuse the same checks from the smith fuzzer. Maybe I should
rename them "source" and "smith" to clarify.

The code remains mostly unchanged, only moved, but I did also remove a
few of the modes. It's not so clear to me any more that it is an
advantage to have a separate typechecking mode when we already have the
evaluation mode. So we can delete some of the code.
