+++ fuzz/src/lib.rs
@@ -6,3 +6,4 @@
 // A copy of the License has been included in the root of the repository.
 
 pub mod smith;
+pub mod uber;
